requirements.txt
Django==1.7.3
psycopg2==2.5.4
ipython==2.3.1
ipdb==0.8
gunicorn==19.1.1
python-social-auth==0.2.1
dj-database-url==0.3.0
django-annoying==0.8.0
pillow==2.7.0
sorl-thumbnail==12.2
